+++ examples/with-thirdpartyemailpassword-passwordless/src/ProtectRoute.tsx
@@ -1,3 +1,4 @@
+import React from "react";
 import { SessionAuth, useSessionContext } from "supertokens-auth-react/recipe/session";
 import { PasswordlessAuth } from "supertokens-auth-react/recipe/passwordless";
 import { ThirdPartyEmailPasswordAuth } from "supertokens-auth-react/recipe/thirdpartyemailpassword";
@@ -10,7 +11,7 @@ import { ThirdPartyEmailPasswordAuth } from "supertokens-auth-react/recipe/third
  * during session creation.
  */
 
-export default function ProtectRoute(props) {
+export default function ProtectRoute(props: any) {
     return (
         <SessionAuth>
             <AuthWrapperSelector {...props} />
@@ -18,7 +19,7 @@ export default function ProtectRoute(props) {
     );
 }
 
-function AuthWrapperSelector(props) {
+function AuthWrapperSelector(props: any) {
     let sessionInfo = useSessionContext();
     if (sessionInfo.accessTokenPayload.isPasswordless === true) {
         return <PasswordlessAuth {...props}>{props.children}</PasswordlessAuth>;