+++ examples/with-thirdpartyemailpassword-passwordless/src/index.tsx
@@ -1,10 +1,11 @@
 import React from "react";
-import { createRoot } from "react-dom/client";
+import ReactDOM from "react-dom/client";
 import "./index.css";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 
-createRoot(document.getElementById("root")).render(
+const root = ReactDOM.createRoot(document.getElementById("root") as HTMLElement);
+root.render(
     <React.StrictMode>
         <App />
     </React.StrictMode>